iommu/arm-smmu-v3: Ensure that a set of associated commands
From 9a1685b3e8a3531121ca33653c10932fce9a95f2 Mon Sep 17 00:00:00 2001
Patch-mainline: Not yet, from openEuler
References: bsn#22
openEuler-commit: 9a1685b3e8a3531121ca33653c10932fce9a95f2

 are inserted in the same ECMDQ

hulk inclusion
category: feature
bugzilla: 174251
CVE: NA

-------------------------------------------------------------------------

The SYNC command only ensures that the command that precedes it in the
same ECMDQ must be executed, but cannot synchronize the commands in other
ECMDQs. If an unmap involves multiple commands, some commands are executed
on one core, and the other commands are executed on another core. In this
case, after the SYNC execution is complete, the execution of all preceded
commands can not be ensured.

Prevent the process that performs a set of associated commands insertion
from being migrated to other cores ensures that all commands are inserted
into the same ECMDQ.

Signed-off-by: Zhen Lei <thunder.leizhen@huawei.com>
Reviewed-by: Hanjun Guo <guohanjun@huawei.com>
Signed-off-by: Zheng Zengkai <zhengzengkai@huawei.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.openEuler/7459-iommu-arm-smmu-v3-Ensure-that-a-set-of-associated-co.patch