page_pool: keep pp info as long as page pool owns the page
From 57f05bc2ab2443b89c2e2562c05053bcc7d30e8b Mon Sep 17 00:00:00 2001
Git-commit: 57f05bc2ab2443b89c2e2562c05053bcc7d30e8b
Patch-mainline: v5.15-rc1
References: bsn#12
openEuler-commit: 156a66dee424e0511725aa8dc52bee49f5096167

Currently, page->pp is cleared and set everytime the page
is recycled, which is unnecessary.

So only set the page->pp when the page is added to the page
pool and only clear it when the page is released from the
page pool.

This is also a preparation to support allocating frag page
in page pool.

Reviewed-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>
Signed-off-by: Yunsheng Lin <linyunsheng@huawei.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.15-rc1-page_pool-keep-pp-info-as-long-as-page-pool-owns-the.patch