xfs: queue inactivation immediately when free realtime
From 65f03d8652b240aa66b99a07e3c423a51e967568 Mon Sep 17 00:00:00 2001
 extents are tight
Git-commit: 65f03d8652b240aa66b99a07e3c423a51e967568
Patch-mainline: v5.15-rc1
References: bsn#12
Modified-by-SEL: No

Now that we have made the inactivation of unlinked inodes a background
task to increase the throughput of file deletions, we need to be a
little more careful about how long of a delay we can tolerate.

Similar to the patch doing this for free space on the data device, if
the file being inactivated is a realtime file and the realtime volume is
running low on free extents, we want to run the worker ASAP so that the
realtime allocator can make better decisions.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>
Reviewed-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.mainline/v5.15-rc1-xfs-queue-inactivation-immediately-when-free-realtim.patch