xfs: fix off-by-one error when the last rt extent is in use
From 9ab72f22277487d2a3ffc8dd20fc918e186bb2b3 Mon Sep 17 00:00:00 2001
Git-commit: 9ab72f22277487d2a3ffc8dd20fc918e186bb2b3
Patch-mainline: v5.15-rc1
References: bsn#12

The fsmap implementation for realtime devices uses the gap between
info->next_daddr and a free rtextent reported by xfs_rtalloc_query_range
to feed userspace fsmap records with an "unknown" owner.  We use this
trick to report to userspace when the last rtextent in the filesystem is
in use by synthesizing a null rmap record starting at the next block
after the query range.

Unfortunately, there's a minor accounting bug in the way that we
construct the null rmap record.  Originally, ahigh.ar_startext contains
the last rtextent for which the user wants records.  It's entirely
possible that number is beyond the end of the rt volume, so the location
synthesized rmap record /must/ be constrained to the minimum of the high
key and the number of extents in the rt volume.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Chandan Babu R <chandanrlinux@gmail.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.15-rc1-xfs-fix-off-by-one-error-when-the-last-rt-extent-is-.patch